These changes are enough to get code to compile, and at least some of the unittests to pass, on FreeBSD 6.0, Solaris 10 x86, and OS X Leopard.  You have to use gmake instead of standard bsd make, though, because neither bsd nor solari make understand 'include */*.d' and 'CFLAGS += $(POPT_FLAGS).'  These may be fixable later.
Most changes fall into four categories:
1) #include differences
2) New errors due to some #ifdef paths being taken differently
3) Undefined functions (see, eg, the new HAVE_STRSEP)
4) Type differences (eg tv_usec is an int on os x, not a long int)

As one concrete example, snprintf.c is an empty file on linux, where
all the functionality is part of glibc.  But on Solaris, some of its
functionality is useful.  This turned up a bug where if you have
vsnprintf on your system but not vasnprintf, then dopr() was not being
used but was being defined, leading to an "unused static function"
warning in gcc.

As another, solaris would complain about "index" being used as a
variable, since it's also a function name.  The var was renamed idx.

Tested by compiling on FreeBSD, Solaris 10, and OS X leopard (x86).  I
also compiled and ran unittests under Linux Ubuntu, to make sure
this change didn't break anything there.

Reviewed by fergus@google.com

+++ trunk/src/exec.c
@@ -455,8 +455,8 @@ int dcc_collect_child(const char *what, pid_t pid,
             rs_log_info("%s times: user %ld.%06lds, system %ld.%06lds, "
                         "%ld minflt, %ld majflt",
                         what,
-                        ru.ru_utime.tv_sec, ru.ru_utime.tv_usec,
-                        ru.ru_stime.tv_sec, ru.ru_stime.tv_usec,
+                        ru.ru_utime.tv_sec, (long) ru.ru_utime.tv_usec,
+                        ru.ru_stime.tv_sec, (long) ru.ru_stime.tv_usec,
                         ru.ru_minflt, ru.ru_majflt);
 
             return 0;
